refactor(web): group browser code by feature
