+++ companion/apps/web/src/wasmJsMain/kotlin/com/rsvpnano/web/BrowserWiring.kt
@@ -2,19 +2,21 @@ package com.rsvpnano.web
 
 import com.rsvpnano.api.ArticleFetchClient
 import com.rsvpnano.api.NanoKtorClient
-import com.rsvpnano.app.FirmwareUpdates
+import com.rsvpnano.updates.FirmwareUpdates
 import com.rsvpnano.app.NanoCompanionController
-import com.rsvpnano.app.NanoEndpoint
-import com.rsvpnano.app.NanoWifiConnector
-import com.rsvpnano.app.NanoWifiEvent
-import com.rsvpnano.app.NanoWifiRequestResult
-import com.rsvpnano.app.NanoWifiSnapshot
-import com.rsvpnano.app.PendingDraftService
+import com.rsvpnano.connection.NanoEndpoint
+import com.rsvpnano.connection.NanoWifiConnector
+import com.rsvpnano.connection.NanoWifiEvent
+import com.rsvpnano.connection.NanoWifiRequestResult
+import com.rsvpnano.connection.NanoWifiSnapshot
+import com.rsvpnano.library.PendingDraftService
 import com.rsvpnano.models.RememberedNano
 import com.rsvpnano.persistence.JsonAppSettingsStore
 import com.rsvpnano.persistence.PendingUploadJsonStore
 import com.rsvpnano.persistence.TextStorage
-import com.rsvpnano.ui.CompanionPresenter
+import com.rsvpnano.presentation.CompanionPresenter
+import com.rsvpnano.web.connection.BrowserNanoApi
+import com.rsvpnano.web.connection.BrowserSerial
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.js.Js
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
