+++ companion/apps/web/src/wasmJsMain/kotlin/com/rsvpnano/web/ui/workspaces/Workspaces.kt
@@ -1,4 +1,4 @@
-package com.rsvpnano.web
+package com.rsvpnano.web.ui.workspaces
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.horizontalScroll
@@ -69,20 +69,21 @@ import com.rsvpnano.models.NanoLocales
 import com.rsvpnano.models.NanoSettings
 import com.rsvpnano.models.NanoSettingsSchema
 import com.rsvpnano.models.PendingUpload
-import com.rsvpnano.models.needsArticleFetch
-import com.rsvpnano.ui.CompanionPresenter
-import com.rsvpnano.ui.CatalogInstall
-import com.rsvpnano.ui.CatalogAsset
-import com.rsvpnano.ui.BookJob
-import com.rsvpnano.ui.CompanionResource
-import com.rsvpnano.ui.CompanionUiState
-import com.rsvpnano.ui.TypographyPreview
-import com.rsvpnano.ui.fontDetails
-import com.rsvpnano.ui.localeDetails
-import com.rsvpnano.ui.byteLabel
-import com.rsvpnano.ui.isArticle
-import com.rsvpnano.ui.librarySubtitle
-import com.rsvpnano.ui.readPercent
+import com.rsvpnano.library.needsArticleFetch
+import com.rsvpnano.presentation.CompanionPresenter
+import com.rsvpnano.presentation.CatalogInstall
+import com.rsvpnano.presentation.CatalogAsset
+import com.rsvpnano.presentation.BookJob
+import com.rsvpnano.presentation.CompanionResource
+import com.rsvpnano.presentation.CompanionUiState
+import com.rsvpnano.ui.settings.TypographyPreview
+import com.rsvpnano.ui.settings.fontDetails
+import com.rsvpnano.ui.settings.localeDetails
+import com.rsvpnano.ui.library.byteLabel
+import com.rsvpnano.ui.library.isArticle
+import com.rsvpnano.ui.library.librarySubtitle
+import com.rsvpnano.ui.library.readPercent
+import com.rsvpnano.web.ui.DetailRow
 import io.github.vinceglb.filekit.dialogs.FileKitType
 import io.github.vinceglb.filekit.dialogs.compose.rememberFilePickerLauncher
 import io.github.vinceglb.filekit.name
