+++ companion/apps/web/src/wasmJsTest/kotlin/com/rsvpnano/web/connection/SerialProtocolTest.kt
@@ -1,5 +1,8 @@
-package com.rsvpnano.web
+package com.rsvpnano.web.connection
 
+import com.rsvpnano.web.connection.SerialFrame
+import com.rsvpnano.web.connection.SerialFrameCodec
+import com.rsvpnano.web.connection.SerialFrameType
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
